remove index, replace it by detail page
