+++ hat/assets/js/apps/Iaso/domains/registry/components/OrgUnitInstances.tsx
@@ -1,37 +1,37 @@
-import React, { FunctionComponent, useEffect, useMemo, useState } from 'react';
+import ErrorOutlineIcon from '@mui/icons-material/ErrorOutline';
+import { Box, Divider, Grid, Paper, Typography } from '@mui/material';
+import { makeStyles } from '@mui/styles';
 import {
-    commonStyles,
     IconButton,
     LoadingSpinner,
+    commonStyles,
     useSafeIntl,
 } from 'bluesquare-components';
-import { makeStyles } from '@mui/styles';
-import { Box, Paper, Typography, Grid, Divider } from '@mui/material';
 import moment from 'moment';
+import React, { FunctionComponent, useEffect, useMemo, useState } from 'react';
 import { useDispatch } from 'react-redux';
-import ErrorOutlineIcon from '@mui/icons-material/ErrorOutline';
 
-import MESSAGES from '../messages';
 import { baseUrls } from '../../../constants/urls';
+import MESSAGES from '../messages';
 
 import { redirectToReplace } from '../../../routing/actions';
-import { useGetEnketoUrl } from '../hooks/useGetEnketoUrl';
 import { useCurrentUser } from '../../../utils/usersUtils';
+import { useGetEnketoUrl } from '../hooks/useGetEnketoUrl';
 
 import InputComponent from '../../../components/forms/InputComponent';
-import InstanceFileContent from '../../instances/components/InstanceFileContent';
 import EnketoIcon from '../../instances/components/EnketoIcon';
+import InstanceFileContent from '../../instances/components/InstanceFileContent';
 
 import { userHasPermission } from '../../users/utils';
 import {
-    useGetOrgUnitInstances,
     useGetInstance,
+    useGetOrgUnitInstances,
 } from '../hooks/useGetInstances';
 
+import * as Permission from '../../../utils/permissions';
+import { LinkToInstance } from '../../instances/components/LinkToInstance';
 import { OrgUnit } from '../../orgUnits/types/orgUnit';
 import { RegistryDetailParams } from '../types';
-import { LinkToInstance } from '../../instances/components/LinkToInstance';
-import * as Permission from '../../../utils/permissions';
 
 type Props = {
     orgUnit: OrgUnit;
@@ -113,7 +113,7 @@ export const OrgUnitInstances: FunctionComponent<Props> = ({
             ...params,
             submissionId,
         };
-        dispatch(redirectToReplace(baseUrls.registryDetail, newParams));
+        dispatch(redirectToReplace(baseUrls.registry, newParams));
     };
     useEffect(() => {
         if (!currentInstanceId && instances && instances?.length > 0) {
