+++ hat/assets/js/apps/Iaso/domains/registry/index.tsx
@@ -1,120 +1,165 @@
-import React, { useState, FunctionComponent } from 'react';
 import { Box, Grid } from '@mui/material';
 import { makeStyles } from '@mui/styles';
 import {
-    TreeViewWithSearch,
-    useSafeIntl,
+    LoadingSpinner,
     commonStyles,
+    useSafeIntl,
 } from 'bluesquare-components';
-
+import { orderBy } from 'lodash';
+import React, { FunctionComponent, useMemo } from 'react';
+import { useDispatch } from 'react-redux';
 import TopBar from '../../components/nav/TopBarComponent';
-import WidgetPaper from '../../components/papers/WidgetPaperComponent';
-import {
-    tooltip,
-    makeTreeviewLabel,
-} from '../orgUnits/components/TreeView/utils';
+import MESSAGES from './messages';
 
-import { OrgUnitLabel, getOrgUnitAncestors } from '../orgUnits/utils';
-import { OrgUnitInfos } from './components/OrgUnitInfos';
+import { getOtChipColors } from '../../constants/chipColors';
+import { baseUrls } from '../../constants/urls';
 
 import {
-    getRootData,
-    getChildrenData,
-    searchOrgUnits,
-} from '../orgUnits/components/TreeView/requests';
+    useGetOrgUnit,
+    useGetOrgUnitListChildren,
+    useGetOrgUnitsMapChildren,
+} from './hooks/useGetOrgUnit';
+
+import { Instances } from './components/Instances';
+import { OrgUnitInstances } from './components/OrgUnitInstances';
+import { OrgUnitPaper } from './components/OrgUnitPaper';
+import { OrgunitTypeRegistry } from './types/orgunitTypes';
+
+import { RegistryDetailParams } from './types';
+
+import { redirectTo } from '../../routing/actions';
+import { OrgUnitTreeviewModal } from '../orgUnits/components/TreeView/OrgUnitTreeviewModal';
+import { OrgUnitBreadcrumbs } from '../orgUnits/components/breadcrumbs/OrgUnitBreadcrumbs';
 import { OrgUnit } from '../orgUnits/types/orgUnit';
 
-import MESSAGES from './messages';
+type Router = {
+    goBack: () => void;
+    params: RegistryDetailParams;
+};
+type Props = {
+    router: Router;
+};
 
 const useStyles = makeStyles(theme => ({
     ...commonStyles(theme),
-    treeContainer: {
-        '& ul[role="tree"]': {
-            maxHeight: '75vh',
-            '& li>div>div>div svg': {
-                display: 'none',
-            },
-        },
-    },
 }));
 
-type Params = {
-    accountId: 'string';
-};
-
-type Props = {
-    params: Params;
-};
-
-export const Registry: FunctionComponent<Props> = ({ params }) => {
+export const Registry: FunctionComponent<Props> = ({ router }) => {
+    const {
+        params: { orgUnitId },
+        params,
+    } = router;
+    const dispatch = useDispatch();
     const classes: Record<string, string> = useStyles();
     const { formatMessage } = useSafeIntl();
 
-    const [selectedOrgUnits, setSelectedOrgUnits] = useState<OrgUnit[]>([]);
+    const { data: orgUnit, isFetching } = useGetOrgUnit(orgUnitId);
+    const { data: orgUnitListChildren, isFetching: isFetchingListChildren } =
+        useGetOrgUnitListChildren(
+            orgUnitId,
+            params,
+            orgUnit?.org_unit_type?.sub_unit_types,
+        );
+    const { data: orgUnitMapChildren, isFetching: isFetchingMapChildren } =
+        useGetOrgUnitsMapChildren(
+            orgUnitId,
+            orgUnit?.org_unit_type?.sub_unit_types,
+        );
+    const subOrgUnitTypes: OrgunitTypeRegistry[] = useMemo(() => {
+        if (!orgUnitMapChildren) {
+            return [];
+        }
+        const options =
+            orgUnit?.org_unit_type?.sub_unit_types.map((subType, index) => ({
+                ...subType,
+                color: getOtChipColors(index) as string,
+                orgUnits: orgUnitMapChildren.filter(
+                    subOrgUnit => subOrgUnit.org_unit_type_id === subType.id,
+                ),
+            })) || [];
+        return orderBy(options, [f => f.depth], ['asc']);
+    }, [orgUnit, orgUnitMapChildren]);
 
-    const [selectedOrgUnitsIds, setSelectedOrgUnitsIds] = useState<string[]>(
-        [],
-    );
-    const [selectedOrgUnitParents, setSelectedOrgUnitParents] = useState<
-        Map<string, string>
-    >(new Map());
-    const onUpdate = (orgUnitIds, parentsData, orgUnits) => {
-        setSelectedOrgUnitsIds(orgUnitIds);
-        setSelectedOrgUnitParents(parentsData);
-        if (orgUnits) {
-            setSelectedOrgUnits(orgUnits);
+    const handleOrgUnitChange = (newOrgUnit: OrgUnit) => {
+        if (newOrgUnit) {
+            const newParams = {
+                ...params,
+                orgUnitId: `${newOrgUnit.id}`,
+            };
+            dispatch(redirectTo(`/${baseUrls.registry}`, newParams));
         }
     };
+
     return (
         <>
-            <TopBar
-                title={formatMessage(MESSAGES.title)}
-                displayBackButton={false}
-            />
-            <Box className={classes.containerFullHeightNoTabPadded}>
+            <TopBar title={formatMessage(MESSAGES.title)} />
+            <Box className={`${classes.containerFullHeightNoTabPadded}`}>
+                {isFetching && <LoadingSpinner />}
+
                 <Grid container spacing={2}>
-                    <Grid item xs={12} md={6} className={classes.treeContainer}>
-                        <TreeViewWithSearch
-                            getChildrenData={id => getChildrenData(id, 'VALID')}
-                            getRootData={(id, type) =>
-                                getRootData(id, type, 'VALID')
-                            }
-                            label={makeTreeviewLabel(classes, true)}
-                            toggleOnLabelClick={false}
-                            request={(value, count) => {
-                                //  @ts-ignore => TODO fix this while migrate tree to Typescript
-                                return searchOrgUnits({ value, count });
-                            }}
-                            makeDropDownText={orgUnit => (
-                                <OrgUnitLabel
+                    {!isFetching && orgUnit && (
+                        <Grid item xs={12}>
+                            <OrgUnitBreadcrumbs
+                                orgUnit={orgUnit}
+                                showRegistry
+                                showOnlyParents
+                            />
+                        </Grid>
+                    )}
+                    <Grid container item xs={12}>
+                        <Grid item xs={4}>
+                            <Box mb={-2}>
+                                <OrgUnitTreeviewModal
+                                    toggleOnLabelClick
+                                    titleMessage={MESSAGES.search}
+                                    onConfirm={handleOrgUnitChange}
+                                    initialSelection={orgUnit}
+                                    defaultOpen={!orgUnitId}
+                                />
+                            </Box>
+                        </Grid>
+                    </Grid>
+                    {!isFetching && orgUnit && (
+                        <>
+                            <Grid item xs={12} md={6}>
+                                <OrgUnitPaper
                                     orgUnit={orgUnit}
-                                    withType
-                                    withSource={false}
+                                    subOrgUnitTypes={subOrgUnitTypes}
+                                    params={params}
+                                    orgUnitListChildren={orgUnitListChildren}
+                                    isFetchingListChildren={
+                                        isFetchingListChildren
+                                    }
+                                    orgUnitMapChildren={orgUnitMapChildren}
+                                    isFetchingMapChildren={
+                                        isFetchingMapChildren
+                                    }
                                 />
-                            )}
-                            toolTip={tooltip}
-                            parseNodeIds={getOrgUnitAncestors}
-                            preselected={selectedOrgUnitsIds}
-                            preexpanded={selectedOrgUnitParents}
-                            selectedData={selectedOrgUnits}
-                            onUpdate={onUpdate}
-                            multiselect={false}
-                            inputLabelObject={MESSAGES.search}
-                        />
-                    </Grid>
-                    <Grid item xs={12} md={6}>
-                        {selectedOrgUnits?.length > 0 && (
-                            <WidgetPaper
-                                title={formatMessage(MESSAGES.selectedOrgUnit)}
+                            </Grid>
+                            <Grid
+                                item
+                                xs={12}
+                                md={6}
+                                alignItems="flex-start"
+                                container
                             >
-                                <OrgUnitInfos
-                                    orgUnit={selectedOrgUnits[0]}
-                                    accountId={params.accountId}
-                                />
-                            </WidgetPaper>
-                        )}
-                    </Grid>
+                                {orgUnit && (
+                                    <OrgUnitInstances
+                                        orgUnit={orgUnit}
+                                        params={params}
+                                    />
+                                )}
+                            </Grid>
+                        </>
+                    )}
                 </Grid>
+                <Box mt={2}>
+                    <Instances
+                        isLoading={isFetching}
+                        subOrgUnitTypes={subOrgUnitTypes}
+                        params={params}
+                    />
+                </Box>
             </Box>
         </>
     );
